ecbuild_generate_yy: remove unused custom target
This commit removes a call to add_custom_target which will define the
same target named 'yy_' on every invocation. This causes name clashes
when ecbuild_generate_yy macro is used more than once in the same
project. I'm not sure about the original intention here, but it seems
save to simply remove this line.