ECBUILD-109 <project>-config.cmake: include TPLs before targets
When using multiple export sets for targets, introduced in ecBuild
6dec8a4, CMake adds a check that any dependent targets from another
export set are present in the automatically generated targets file.

For example, in atlas-targets.cmake, a check for the presence of the
targets eckit_geometry and eckit_mpi is added.

Therefore it is crucial to include all TPLs and their targets before
including the targets of the current package.